pyproject.toml
[project]
name = "drf-boilerplate"
version = "0.1.0"
description = "Django REST API template with essential features and best practices."
authors = [
    { name = "Full Name", email = "dev@email.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "babel>=2.17.0",
    "cffi>=1.17.1",
    "colorlog>=6.10.1",
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "django-unfold>=0.63.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "markdown>=3.8.2",
    "psycopg2-binary>=2.9.10",
    "pycparser>=2.22",
    "pygments>=2.19.2",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "ruff>=0.15.0",
    "sentry-sdk>=2.34.1",
    "sqlparse>=0.5.3",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "django-query-counter>=0.4.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]

[tool.black]
line-length = 88
target-version = ["py313"]
extend-exclude = '''
/(
  migrations
)/
'''

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src", "src/apps"]
extend-exclude = [
    "*/migrations/*",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
]

[tool.ruff.lint.isort]
known-first-party = [
    "account",
    "api",
    "attendance",
    "bike",
    "bot",
    "config",
    "core",
    "gamification",
    "payroll",
    "rules",
    "ticket",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
pythonpath = ["src", "src/apps"]
testpaths = ["src/tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov --cov-report=term-missing --cov-report=xml --cov-report=html"

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/migrations/*",
    "*/tests/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
precision = 2
